# Optimize pack()