--- a/ext/standard/pack.c
+++ b/ext/standard/pack.c
@@ -76,10 +76,7 @@ ZEND_ATTRIBUTE_CONST static inline uint16_t php_pack_reverse_int16(uint16_t arg)
 /* {{{ php_pack_reverse_int32 */
 ZEND_ATTRIBUTE_CONST static inline uint32_t php_pack_reverse_int32(uint32_t arg)
 {
-	uint32_t result;
-	result = ((arg & 0xFF) << 24) | ((arg & 0xFF00) << 8) | ((arg >> 8) & 0xFF00) | ((arg >> 24) & 0xFF);
-
-	return result;
+	return ZEND_BYTES_SWAP32(arg);
 }
 /* }}} */