@nielsdos 2025-06-08
> maybe a interessting benchmark/test-case for `unpack`: [stomp-php/stomp-php#184](https://github.com/stomp-php/stomp-php/pull/184)
> 
> ```
> <?php
> $file = file_get_contents('FILE');
> echo count(unpack('C*', $file)) . "\n";
> ```
> 
> vs.
> 
> ```
> <?php
> $file = file_get_contents('FILE');
> echo strlen($file) . "\n";
> ```
> 
> using `truncate -s 80M FILE`.
> 
> the `strlen()` variant is a lot faster

Unpack will always be slower than just strlen. However, your code revealed that repetitions were handled in a slow way where lots of temporary strings were created and then parsed. I opened a PR to fix that particular issue: https://github.com/php/php-src/pull/18803