@divinity76 2025-06-10
could benchmark
```
static inline void php_pack(const zval *val, size_t size,
                            php_pack_endianness enc, char *out)
{
    zend_long z = zval_get_long(val);

    if ((enc == PHP_LITTLE_ENDIAN) != MACHINE_LITTLE_ENDIAN) {
        z = PHP_LONG_BSWAP(z);
    }
    memcpy(out, (char*)&z + sizeof(z) - size, size);
}
```
might be faster