@nielsdos 2025-06-10
I think I managed to make the compiler happy and let it make good inlining decisions while keeping the code simple.

For example for this:
```php
for ($i = 0; $i < 10_000_000; ++$i) {
  pack("J", 0x7FFFFFFFFFFFFFFF);
}
```

On an i7-4790:
```
Benchmark 1: ./sapi/cli/php pack.php
  Time (mean ± σ):     408.8 ms ±   3.4 ms    [User: 406.1 ms, System: 1.6 ms]
  Range (min … max):   403.6 ms … 413.6 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old pack.php
  Time (mean ± σ):     451.7 ms ±   7.7 ms    [User: 448.5 ms, System: 2.0 ms]
  Range (min … max):   442.8 ms … 461.2 ms    10 runs
 
Summary
  ./sapi/cli/php pack.php ran
    1.11 ± 0.02 times faster than ./sapi/cli/php_old pack.php
```

And for this:
```php
for ($i=0;$i<4000000;$i++)
pack("nvc*", 0x1234, 0x5678, 65, 66);
```

On the same machine:
```
Benchmark 1: ./sapi/cli/php pack.php
  Time (mean ± σ):     239.3 ms ±   6.0 ms    [User: 236.2 ms, System: 2.3 ms]
  Range (min … max):   233.2 ms … 256.8 ms    12 runs
 
Benchmark 2: ./sapi/cli/php_old pack.php
  Time (mean ± σ):     271.9 ms ±   3.3 ms    [User: 269.7 ms, System: 1.3 ms]
  Range (min … max):   267.4 ms … 279.0 ms    11 runs
 
Summary
  ./sapi/cli/php pack.php ran
    1.14 ± 0.03 times faster than ./sapi/cli/php_old pack.php
```

Let's hope it's reproducible