@divinity76 2025-06-15
If someone has a big endian server running ssh they'd be willing to share, that would be helpful.
Fwiw nielsdos ssh keys are available at https://github.com/nielsdos.keys

Likewise my keys at https://github.com/divinity76.keys