@bukka 2025-06-15
You should be able to emulate in qemu but haven't tried it myself so not sure if it's any good.