@divinity76 2025-06-18
as for the question "is it (still) faster?",
yes, on @NattyNarwhal 's BE it is faster:
```
divinity@php-ci-ppc64be ~/php-src $ cat bench.php 
<?php
declare(strict_types=1);
for ($i=0;$i<4000000;$i++)
pack("nvc*", 0x1234, 0x5678, 65, 66);
divinity@php-ci-ppc64be ~/php-src $ hyperfine --warmup 3 --runs 20 \
  './sapi/cli/php ./bench.php' \
  './sapi/cli/php_old ./bench.php'
Benchmark 1: ./sapi/cli/php ./bench.php
  Time (mean ± σ):      1.059 s ±  0.038 s    [User: 1.053 s, System: 0.005 s]
  Range (min … max):    1.033 s …  1.170 s    20 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 2: ./sapi/cli/php_old ./bench.php
  Time (mean ± σ):      1.119 s ±  0.006 s    [User: 1.114 s, System: 0.005 s]
  Range (min … max):    1.111 s …  1.131 s    20 runs
 
Summary
  ./sapi/cli/php ./bench.php ran
    1.06 ± 0.04 times faster than ./sapi/cli/php_old ./bench.php

```